==> LPS/analysis/3-differential.Rmd <==
---
title: "DS Analysis"
author: 
- name: Helena L. Crowell
  affiliation:
  - &IMLS Institute for Molecular Life Sciences, University of Zurich, Switzerland
  - &SIB Swiss Institute of Bioinformatics (SIB), University of Zurich, Switzerland  
- name: Pierre-Luc Germain
  affiliation: 
  - *IMLS
  - *SIB
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output:
  html_document:
    toc: true
---

```{r include = FALSE}
knitr::opts_chunk$set(cache = TRUE, autodep = TRUE)
```

# Load packages

```{r load-libs, message = FALSE, warning = FALSE}
library(dplyr)
library(muscat)
library(scater)
library(SingleCellExperiment)
library(UpSetR)
```

# Load data

```{r load-data}
sce <- readRDS(file.path("output", "3-SCE_clustered.rds"))

# make WT reference group
sce$group_id <- relevel(sce$group_id, ref = "WT")

# reorder sample levels
m <- match(levels(sce$sample_id), sce$sample_id)
o <- order(sce$group_id[m])
sce$sample_id <- factor(sce$sample_id, levels = levels(sce$sample_id)[o])

# prep. SCE for 'muscat'
sce <- prepSCE(sce,
  cluster_id = "cluster_id",
  sample_id = "sample_id",
  group_id = "group_id",
  drop = FALSE)
```

# Aggregation

```{r pb, message = FALSE, fig.width = 6}
# compute pseudobulk (sum of counts) by cluster-sample
(pb <- aggregateData(sce, assay = "counts", 
    by = c("cluster_id", "sample_id"), fun = "sum"))

# pseudobulk-level MDS plot
pbMDS(pb)
```

# DS analysis

```{r ds}
# run edgeR on pseudobulks
res <- pbDS(pb, method = "edgeR", verbose = FALSE)

# reformat results
tbl <- resDS(sce, res, bind = "col")

# filter & sort
tbl <- filter(tbl, p_adj.loc < 0.05, logFC > 1)
tbl <- arrange(tbl, p_adj.loc)

# no. of DS genes per cluster
res_by_k <- split(tbl, tbl$cluster_id)
vapply(res_by_k, nrow, numeric(1))

# top hits in ea. cluster
top <- do.call("rbind", lapply(res_by_k, head, 3))
top <- select(top, -c("contrast", "p_adj.glb", "F"))
top$gene <- gsub("^.*\\.", "", top$gene)
format(data.frame(top, row.names = NULL), digits = 3)
```

# Pseudobulk-level heatmap

```{r fig.width = 8, fig.height = 5}
# single gene across all clusters
pbHeatmap(sce, res, g = "ENSMUSG00000035385.Ccl2")

# top-20 in single cluster
pbHeatmap(sce, res, k = "Astrocytes")

# top-5 in ea. cluster
pbHeatmap(sce, res, top_n = 5)
```

# Violin plots

```{r fig.width = 8, fig.height = 5}
# split cells by cluster
cs_by_k <- split(colnames(sce), sce$cluster_id)
plotExpression(sce[, cs_by_k$Endothelial], 
    features = res_by_k$Endothelial$gene[seq_len(8)],
    x = "sample_id", colour_by = "group_id") + theme_classic() + 
    theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

# Between-cluster concordance

```{r fig.width = 8, fig.height = 5}
ds_gs <- lapply(res_by_k, pull, "gene")
upset(fromList(ds_gs), sets = levels(sce$cluster_id))
```


# Write results to .rds

```{r save}
saveRDS(res, file.path("output", "DS_results.rds"))
```
